The truncation fix was not enough! There will still be a truncation when calling the node insert function. Change the length types to size_t to avoid length truncation when calling insert functions.

+++ src/jes_tree.c
@@ -167,7 +167,7 @@ struct jes_node* jes_tree_get_container_parent_node(struct jes_context* ctx,
 
 struct jes_node* jes_tree_insert_node(struct jes_context* ctx,
                                       struct jes_node* parent, struct jes_node* anchor,
-                                      uint16_t type, uint16_t length, const char* value)
+                                      uint16_t type, size_t length, const char* value)
 {
   struct jes_node *new_node = NULL;
 
@@ -359,7 +359,7 @@ void jes_tree_delete_node(struct jes_context* ctx, struct jes_node* node)
 struct jes_node* jes_tree_insert_key_node(struct jes_context* ctx,
                                           struct jes_node* parent_object,
                                           struct jes_node* anchor,
-                                          uint16_t keyword_length,
+                                          size_t keyword_length,
                                           const char* keyword)
 {
   struct jes_node* new_node = NULL;
